harden: Pass 4 — 17 fixes across 11 files (final production sweep)
HIGH (2):
- llm.c/llm.h: vocab hash int16_t→int32_t (LLM_MAX_VOCAB=160000 overflows int16)
- tensor_ipc.c: IPC channel slot reuse — scan destroyed slots before growing pool

MEDIUM (13):
- tensor_ipc.c: NULL check after tensor_alloc_shared(), pipeline count<2 guard
- virt.c: container slot reuse + shared memory freed on destroy
- netstack.c: 503 path now closes TCP connection; API key constant-time compare
- ssh.c: channel window decrements by bytes buffered not received; signed_data 512→768
- train_service.c: cosine LR uses Bhaskara I (~0.1% error) replacing broken Taylor
- git.c: git_ref_resolve bounded recursion (depth limit 10)
- deploy_service.c: scale-down now destroys removed replicas' MEUs
- llm.c: sampling uses crypto_random() instead of TSC

LOW (2):
- tensor_monitor.c: check_alerts returns early when sample_count==0
- aishell.c: shell_atoi overflow guard

Build: 48 sources, 0 errors, 0 warnings, 631896 bytes
Tests: 16/16 selftests PASS, Braniac PASS

+++ runtime/nn/llm.h
@@ -107,7 +107,7 @@ typedef struct {
     float             *vocab_scores; /* [vocab_size] merge scores */
 
     /* Hash table for O(1) token lookup by string */
-    int16_t           *vocab_ht_slot; /* [LLM_HASH_SIZE] → vocab index or -1 */
+    int32_t           *vocab_ht_slot; /* [LLM_HASH_SIZE] → vocab index or -1 */
 
     /* Model data buffer */
     void    *data_buf;          /* Loaded GGUF file */
